requirements.txt
fastapi
uvicorn
requests
loguru
python-dotenv
textual
textual-plotext
rich
stravalib
